requirements.txt
Django==4.2.7
python-dotenv==1.0.0
requests>=2.31.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
djangorestframework-simplejwt>=5.3.0
beautifulsoup4>=4.12.2
langchain-google-genai
langchain
duckduckgo-search
langchain-community
whitenoise>=6.6.0
psycopg2-binary>=2.9.0
dj-database-url>=2.1.0